feat!: reimplement nodeset controller
The original NodeSet controller implementation was based on
StatefulSet+DaemonSet. DaemonSet does not support replicas without
smart meta scheduling (better than the pod scheduler), but would always
inherantly be inefficient and behind the pod scheduler.

This new implementation is based on StatefulSet with unordered scale-in
and is fully parallel (never monotonic).

This new implementation should be in parity with the old version.