feat: deduplicate stdout/logfile streams
slurmctld and slurmdbd do not need to log anything, because its stdout
already contains everything its log would, but without a timestamp.
Using `LogTimeFormat=*,format_stderr` makes stdout match the logfile
structure.

The slurmd logfile is needed to capture the slurmstepd output, which
runs as a detached/orphaned process but logs to SlurmdLogFile. Using a
socket as the logfile and a side car container to read the socket, we
can capture and propagate the slurmd+slurmstepd logs as a separate
stdout stream.