feat: convert reconfigure from side-car to job
We only need to reconfigure when helm does an upgrade to the chart. At
that time, the config files may have changed but the slurmd's will not
automatically reconfigure until commanded to.

This also stops a shutdown race condition between the update to the
slurm.conf in configmap and issuing a reconfigure when the slurmctld is
handling SIGTERM/SIGQUIT on pod termination. This can lead to the helm
chart failing unintentionally.