feat: make accounting.external link to external database
Previously, accounting.external meant to point to an existing slurmdbd.
This configuration, while valid, was broken in the chart. Additionally,
a subchart to spin up just a slurmdbd, or slurmdbd with database does
not exist, so its value was limited.

This change allows you to use the slurm chart with an externally managed
database. Be it a cloud managed database or one created from chart.