feat: remove bitnami/mariadb subchart
We do not want to control nor prescribe what database to use and how
you should deploy it, hence the removal of the subchart.

There are multiple implementations which provide a mysql/mariadb
database for Slurm. Their features, support, and ergonomics vary.
Consider the mariadb-operator/mariadb-operator [1].

The slurmctld persistence was enabled because it is a better default for
Slurm. It was disabled previousely because bitnami/mariadb has its
persistence disabled by default. The addition of a Slurm cluster SLUID
would cause issues with cluster, so ensuring disabling persistence was a
easier workaround.

[1]: https://github.com/mariadb-operator/mariadb-operator