feat: remove most initconf initContainer
Slurm configuration files are directly volume mounted with pod
`securityGroup.fsGroup`. This means that the initConf container is no
longer required for Slurm components, except for slurmdbd. It does not
allow 0o640 permissions on slurmdbd.conf.

The reconfigure script was rewritten with pure coreutils (no rsync).
Because we are directly mounting the Slurm config files, cannot rely on
rsync to determine when files have changed. Instead we can hash the
files in the volume mounted directory.