feat: create controller to manage slurm clients
The Cluster CRD originally handled this function, then moved into the
new Controller CRD, now into a new internal controller.

The creation of this controller allows the slurm client management to
isolated. Although, it still is highly dependant on the Controller CRD
because that is the key for the client map.