fix: refresh pod status before possible drain request
The pod status may be stale at this point so we need to refresh before
taking action based on status. Additionally, we only really care if the
pod is running or not. It's ready status does not matter to Slurm, hence
the NodeSet controller should not consider ready status for this logic.