feat: add Token CRD
The Token CR defines a JWT secret to be created. If allowed, the
operator will refresh the token such that other services will have a
valid token that changes over time. Otherwise an immutable token is
created.

This is useful for other services that need to communicate with Slurm
via slurmrestd (e.g. slurm-bridge, slurm-exporter, etc.).