chore: add liveness probe to slurmctld container
Kubernetes does not currently have visibility into the state of the
slurmctld pod on an ongoing basis, and as such is not able to detect
failures of the slurmctld process within this pod. Adding a liveliness
probe to the slurmctld pod shuold make it such that Kubernetes can
detect and remedy failues of this process.

Ticket: 23375