fix: set imagePullPolicy in slurm-operator chart from values
imagePullPolicy for the operator and webhook pods in the slurm-operator
Helm chart were being defaulted to IfNotPresent regardless of the values
set in values.yaml. This commit matches the behavior used in the Slurm
Helm chart, whereby component imagePullPolicies default to a global
imagePullPolicy if not set in each component.

Changelog: fixed - Fixed imagePullPolicy in slurm-operator Helm chart