fix: update NodeSet pod state on pod migration
There is an edge case where a NodeSet pod may not de-register when
slurmctld goes down and the pod lands on a different kube node. If this
occurs, the pod inherits the Slurm node state but said state is invaild
given the different Kube node. Slurm-operator shall reset the state of
Slurm nodes who are still registered but changed Kube node.

Changelog: fixed - Fixes edge case where Slurm node state is not reset
 when a worker pod migrates kube nodes